% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{detour_index_sf}
\alias{detour_index_sf}
\title{detour_index() for every row of an sf data frame}
\usage{
detour_index_sf(x)
}
\arguments{
\item{x}{an sf data frame}
}
\value{
\code{x} with one new column, detour_index
}
\description{
detour_index() for every row of an sf data frame
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
res <- detour_index_sf(nc[1:5, ])
res$detour_index
}
