% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{exchange_index_sf}
\alias{exchange_index_sf}
\title{exchange_index() for every row of an sf data frame}
\usage{
exchange_index_sf(x)
}
\arguments{
\item{x}{an sf data frame}
}
\value{
\code{x} with one new column, exchange_index
}
\description{
exchange_index() for every row of an sf data frame
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
res <- exchange_index_sf(nc[1:5, ])
res$exchange_index
}
