% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{hull_ratio_index}
\alias{hull_ratio_index}
\title{Convex hull area ratio of a (multi)polygon}
\usage{
hull_ratio_index(poly)
}
\arguments{
\item{poly}{a single sfg/sfc (MULTI)POLYGON}
}
\value{
list(index, area, hull_area, hull). index is in (0, 1], 1 = the
polygon is convex (equals its own convex hull).
}
\description{
Convex hull area ratio of a (multi)polygon
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
hull_ratio_index(nc[nc$NAME == "Wake", ])$index
hull_ratio_index(nc[nc$NAME == "Dare", ])$index
}
