% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{hull_ratio_index_sf}
\alias{hull_ratio_index_sf}
\title{hull_ratio_index() for every row of an sf data frame}
\usage{
hull_ratio_index_sf(x)
}
\arguments{
\item{x}{an sf data frame}
}
\value{
\code{x} with one new column, hull_ratio_index
}
\description{
Each row indexed independently (see shape_indices_sf() for a version
that also computes the other indices).
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
res <- hull_ratio_index_sf(nc[1:5, ])
res$hull_ratio_index
}
