% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{polsby_popper_index}
\alias{polsby_popper_index}
\title{Polsby-Popper compactness score of a (multi)polygon}
\usage{
polsby_popper_index(poly)
}
\arguments{
\item{poly}{a single sfg/sfc (MULTI)POLYGON}
}
\value{
list(index, area, perimeter)
}
\description{
4\emph{pi}area/perimeter^2, in (0, 1], 1 = a circle. Perimeter is the full
boundary length (outer ring plus any holes).
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
polsby_popper_index(nc[nc$NAME == "Wake", ])$index
polsby_popper_index(nc[nc$NAME == "Dare", ])$index
}
