% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{polsby_popper_index_sf}
\alias{polsby_popper_index_sf}
\title{polsby_popper_index() for every row of an sf data frame}
\usage{
polsby_popper_index_sf(x)
}
\arguments{
\item{x}{an sf data frame}
}
\value{
\code{x} with one new column, polsby_popper_index
}
\description{
polsby_popper_index() for every row of an sf data frame
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
res <- polsby_popper_index_sf(nc[1:5, ])
res$polsby_popper_index
}
