% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{reock_index}
\alias{reock_index}
\title{Reock compactness score of a (multi)polygon}
\usage{
reock_index(poly)
}
\arguments{
\item{poly}{a single sfg/sfc (MULTI)POLYGON}
}
\value{
list(index, area, mbc_area, mbc). \code{mbc} is the minimum bounding
circle itself, as an sfc POLYGON, for plotting.
}
\description{
area/area(minimum bounding circle), in (0, 1], 1 = a circle.
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
reock_index(nc[nc$NAME == "Wake", ])$index
reock_index(nc[nc$NAME == "Dare", ])$index
}
