% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{reock_index_sf}
\alias{reock_index_sf}
\title{reock_index() for every row of an sf data frame}
\usage{
reock_index_sf(x)
}
\arguments{
\item{x}{an sf data frame}
}
\value{
\code{x} with one new column, reock_index
}
\description{
reock_index() for every row of an sf data frame
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
res <- reock_index_sf(nc[1:5, ])
res$reock_index
}
