% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeindices-package.R
\docType{package}
\name{shapeindices-package}
\alias{shapeindices}
\alias{shapeindices-package}
\title{shapeindices: Shape Indices for (Multi)Polygons}
\description{
Convexity, moment-of-inertia compactness, span (mean-pairwise-distance),
and radial concentration (mean-distance-to-geometric-median) indices
for (multi)polygons via constrained Delaunay triangulation (CDT), plus
classic boundary/hull/bounding-box compactness scores from the
redistricting literature (convex-hull ratio, Polsby-Popper,
width-length ratio, Reock) that need no triangulation at all. See
\code{vignette("c-understanding-convexity-index", package = "shapeindices")} for the full derivation
and worked examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nkaza/shapeindices}
  \item \url{https://nkaza.github.io/shapeindices/}
  \item Report bugs at \url{https://github.com/nkaza/shapeindices/issues}
}

}
\author{
\strong{Maintainer}: Nikhil Kaza \email{nkaza@unc.edu}

Authors:
\itemize{
  \item Nikhil Kaza \email{nkaza@unc.edu}
}

}
\keyword{internal}
