% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-clipping.R
\name{tri_quad_points}
\alias{tri_quad_points}
\title{3-point interior (Hammer-Stroud) quadrature rule for a triangle}
\usage{
tri_quad_points(coords3x2)
}
\arguments{
\item{coords3x2}{a 3x2 matrix of the triangle's corner coordinates}
}
\value{
a 3x2 matrix of quadrature point coordinates
}
\description{
3-point interior (Hammer-Stroud) quadrature rule for a triangle
}
\examples{
triangle <- rbind(c(0, 0), c(4, 0), c(0, 3))
tri_quad_points(triangle)
}
