% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical-metrics.R
\name{width_length_ratio_index_sf}
\alias{width_length_ratio_index_sf}
\title{width_length_ratio_index() for every row of an sf data frame}
\usage{
width_length_ratio_index_sf(x)
}
\arguments{
\item{x}{an sf data frame}
}
\value{
\code{x} with one new column, width_length_ratio_index
}
\description{
width_length_ratio_index() for every row of an sf data frame
}
\examples{
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
res <- width_length_ratio_index_sf(nc[1:5, ])
res$width_length_ratio_index
}
